Added tests for enum conversion
