+++ src/emulator/storage/upload.ts
@@ -248,3 +248,19 @@ export class UploadService {
     return encodeURIComponent(`${uploadId}_b_${bucketId}_o_${objectId}`);
   }
 }
+
+/**
+ *  Converts an UploadStatus enum to a string.
+ */
+export function uploadStatusEnumToStr(status: UploadStatus) {
+  switch (status) {
+    case UploadStatus.ACTIVE:
+      return "active";
+    case UploadStatus.CANCELLED:
+      return "cancelled";
+    case UploadStatus.FINISHED:
+      return "final";
+    default:
+      return "unknown";
+  }
+}
